Adds less important files that end up missing from the project.
These files should be installed on the process of NuGet's
installation. But during restore, no NuGet content is restored
on the working project paths.

So, to avoid having missing files from the project after a first
build on a fresh repo clone, these files were added.

They are not important and would not interfere on the working of
the project. They are here just because they are added to the
project by default when the NuGet package is installed.