Fixes .gitattributes to ensure LF ending on files.
Adding 'text eol=lf' forces the file extension to be treated as a text
file and also ensures it will be checked out in that line ending
regardless of the git client/os being used.

This commit adds the .sh extension and updates for several other
existing ones in .gitattributes.